# specs/openapi.generated.yaml
openapi: 3.1.0
info:
  title: OverVault API
  version: 0.1.0
paths:
...
  /api/auth/dev-login:
    post:
      tags:
      - auth
      summary: Dev Login
      operationId: dev_login_api_auth_dev_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - HTTPBearer: []
  /api/auth/nonce:
    post:
      tags:
      - auth
      summary: Wallet Nonce
      operationId: wallet_nonce_api_auth_nonce_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletNonceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletNonceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/wallet-login:
    post:
      tags:
      - auth
      summary: Wallet Verify
      operationId: wallet_verify_api_auth_wallet_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/files:
    get:
      tags:
      - files
      summary: List Files
      operationId: list_files_api_files_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FileOut'
                type: array
                title: Response List Files Api Files Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - files
      summary: Upload File
      operationId: upload_file_api_files_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_api_files_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/files/{file_id}:
    get:
      tags:
      - files
      summary: Get File
      operationId: get_file_api_files__file_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - files
      summary: Delete File
      operationId: delete_file_api_files__file_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/{file_id}/download:
    get:
      tags:
      - files
      summary: Download Current
      operationId: download_current_api_files__file_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/{file_id}/verify:
    post:
      tags:
      - files
      summary: Verify File
      description: 'Matches the mock''s POST /files/:id/verify. Recomputes and compares
        the

        local hash on every call; chain_hash mirrors it until Sriganesh''s ChainService

        is wired in (see ADR 0005) - at that point chain_hash comes from MSTScan and

        `verified` reflects an actual on-chain comparison.'
      operationId: verify_file_api_files__file_id__verify_post
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/files/{file_id}/protection:
    put:
      tags:
      - files
      summary: Set Protection
      operationId: set_protection_api_files__file_id__protection_put
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtectionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_file_api_files_post:
      properties:
        upload:
          type: string
          contentMediaType: application/octet-stream
          title: Upload
        comment:
          type: string
          title: Comment
          default: ''
      type: object
      required:
      - upload
      title: Body_upload_file_api_files_post
...
    DevLoginRequest:
      properties:
        email:
          type: string
          title: Email
      type: object
      required:
      - email
      title: DevLoginRequest
    FileOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        owner:
          type: string
          title: Owner
        size:
          type: integer
          title: Size
        protection:
          $ref: '#/components/schemas/ProtectionMode'
        verification:
          type: string
          title: Verification
        hash:
          type: string
          title: Hash
        ownership_tx:
          anyOf:
          - type: string
          - type: 'null'
          title: Ownership Tx
        content_type:
          type: string
          title: Content Type
        current_version:
          type: integer
          title: Current Version
        approved_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Approved Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        my_access:
          anyOf:
          - $ref: '#/components/schemas/PermissionLevel'
          - type: 'null'
      type: object
      required:
      - id
      - name
      - owner
      - size
      - protection
      - verification
      - hash
      - content_type
      - current_version
      - created_at
      - updated_at
      title: FileOut
      description: 'Field names/shape match Pavan''s frontend mock (handlers.ts) exactly
        for the

        top-level fields; extra fields are additive so nothing he reads breaks.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PermissionLevel:
      type: string
      enum:
      - read
      - write
      - manage
      title: PermissionLevel
...
    ProtectionMode:
      type: string
      enum:
      - none
      - append_only
      - read_only
      title: ProtectionMode
    ProtectionUpdate:
      properties:
        protection:
          $ref: '#/components/schemas/ProtectionMode'
      type: object
      required:
      - protection
      title: ProtectionUpdate
    Role:
      type: string
      enum:
      - employee
      - manager
      - admin
      - auditor
      title: Role
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - user
      title: TokenResponse
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        name:
          type: string
          title: Name
        role:
          $ref: '#/components/schemas/Role'
        wallet_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Address
      type: object
      required:
      - id
      - email
      - name
      - role
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyResult:
      properties:
        file_id:
          type: string
          title: File Id
        local_hash:
          type: string
          title: Local Hash
        chain_hash:
          type: string
          title: Chain Hash
        verified:
          type: boolean
          title: Verified
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - file_id
      - local_hash
      - chain_hash
      - verified
      - timestamp
      title: VerifyResult
      description: 'Matches POST /files/{id}/verify in the mock. Until Sriganesh''s
        ChainService

        is wired in, chain_hash mirrors local_hash and verified is always true.'
...
    WalletNonceRequest:
      properties:
        address:
          type: string
          title: Address
      type: object
      required:
      - address
      title: WalletNonceRequest
    WalletNonceResponse:
      properties:
        nonce:
          type: string
          title: Nonce
        message:
          type: string
          title: Message
      type: object
      required:
      - nonce
      - message
      title: WalletNonceResponse
    WalletVerifyRequest:
      properties:
        address:
          type: string
          title: Address
        nonce:
          type: string
          title: Nonce
        signature:
          type: string
          title: Signature
      type: object
      required:
      - address
      - nonce
      - signature
      title: WalletVerifyRequest
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer